Added Resident macrophage population
Resident macrophages are the ones already in the LP since embryonic origin. These have been shown to proliferate in presence of HP there by creating a regulatory micro-environment. They release IL10 and TGF-beta. 
The other changes are in reference to the introduction of this new population. The regulatory and inflammatory macrophages are both from bone derived macrophages and hence are obtained after state change in monocytes.